# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Waves API
  version: 1.0.0
  summary: Read and write a Waves shared ledger — groups, expenses, settlements and balances.
  description: |
    The Waves API lets a program act for a Waves user: read the groups they are
    in, add and edit expenses, record payments, and read what everybody is owed.
    Every request runs as that person with their row-level security intact, so a
    token can never reach a group its owner could not reach from the app. A scope
    only narrows what the token may do; it never widens what the person can see.

    ## Authenticating

    There are two ways in, and they produce the same kind of credential.

    **A personal access token** is a developer acting on their own account. It is
    created in the Waves developer console, carries a fixed scope list, and is
    sent as `Authorization: Bearer wavs_pat_…`. Use this when the program and the
    account belong to the same person.

    **OAuth 2.1 authorization code with PKCE** is somebody else's program acting
    for a Waves user who approved it. Send the browser to `/oauth/authorize`,
    exchange the returned code at `/oauth/token`, and use the resulting
    `wavs_at_…` access token in the same `Authorization` header. `S256` is the
    only code challenge method accepted, and a refresh token is only issued when
    `offline_access` was granted. Refresh tokens rotate: each one is single-use,
    and replaying a spent one disconnects the application.

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/groups:
    get:
      tags: [Groups]
      operationId: listGroups
      summary: List the groups this person is in
      description: >-
        Newest first. Archived groups are left out unless `include_archived` is
        the literal string `true`.
      security:
        - bearerAuth: []
        - oauth2: [groups.read]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: include_archived
          in: query
          required: false
          description: Include archived groups. Defaults to false.
          schema:
            $ref: '#/components/schemas/QueryFlag'
      responses:
        '200':
          description: One page of groups.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupPage'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [Groups]
      operationId: createGroup
      summary: Create a group
      description: >-
        The caller becomes its first member and an admin. The group's id is
        derived from the `Idempotency-Key`, so a retried create returns the same
        group rather than a second one.
      security:
        - bearerAuth: []
        - oauth2: [groups.write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: The group as created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /v1/groups/{groupId}/members:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    get:
      tags: [Groups]
      operationId: listGroupMembers
      summary: Who is in a group
      description: >-
        Returns the complete list. Not paginated — a group's membership is
        bounded — so the answer is a plain `{"data": […]}` with no cursor and
        there is nothing to loop over. Oldest first.
      security:
        - bearerAuth: []
        - oauth2: [groups.read]
      parameters:
        - name: include_left
          in: query
          required: false
          description: >-
            Include people who have left. Their history and balances count
            either way; they just stop accruing new shares. Defaults to false.
          schema:
            $ref: '#/components/schemas/QueryFlag'
      responses:
        '200':
          description: The group's members.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberList'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [Groups]
      operationId: addGroupMember
      summary: Add a person to a group
      description: >-
        Adds a *ghost*: a name in the group, with expenses filed against it,
        that a real person can later claim through an invite. There is no way to
        add somebody else's account to a group from the API, and there should
        not be — joining is something a person does, not something done to them.
        The member's id is derived from the `Idempotency-Key`.
      security:
        - bearerAuth: []
        - oauth2: [groups.write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
      responses:
        '201':
          description: The member as added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /v1/friends:
    get:
      tags: [Friends]
      operationId: listFriends
      summary: Who you owe and who owes you
      description: >-
        Not a friend list — there is no such table and nothing to add somebody
        to. It is the derived answer to "who am I not square with", netted
        across every group you share with them, per currency. That is why there
        is no write here: adding a person is creating a group with them in it,
        which is `POST /v1/groups` followed by
        `POST /v1/groups/{groupId}/members`. Returns the complete list — not
        paginated, nothing to loop over.
      security:
        - bearerAuth: []
        - oauth2: [friends.read]
      responses:
        '200':
          description: Everybody you are not square with.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendList'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ───────────────────────────────────────────────────────── categories ──
  /v1/categories:
    get:
      tags: [Categories]
      operationId: listCategories
      summary: This person's expense categories
      description: >-
        The caller's own tag catalogue: rows that either override a built-in
        category's presentation or define a custom tag of their own. Deleted
        rows are left out. Ordered by `sort_order`. Returns the complete list —
        not paginated, nothing to loop over.
      security:
        - bearerAuth: []
        - oauth2: [categories.read]
      responses:
        '200':
          description: The catalogue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryList'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [Categories]
      operationId: createCategory
      summary: Add a category, or override a built-in one
      description: >-
        Two shapes, and exactly one of them per call. Send `label` (and
        optionally `icon` and `tint`) to mint a custom tag; send `builtin_id` to
        override a built-in category's presentation. Sending both means you have
        not decided which you meant, and is refused.


        The row id is derived from `Idempotency-Key` like every other create
        here, so a retry collides with the row it already wrote and comes back
        409 rather than adding a second tag to somebody's picker. With no key
        the id is random, which is what sending no key means.
      security:
        - bearerAuth: []
        - oauth2: [categories.write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: The category as created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/categories/{categoryId}:
    parameters:
      - $ref: '#/components/parameters/CategoryId'
    patch:
      tags: [Categories]
      operationId: updateCategory
      summary: Change a category
      description: >-
        Send at least one of `label`, `icon`, `tint`, `hidden` or `sort_order`.
        Hiding a built-in override is how a person removes a category they never
        use without deleting anything.
      security:
        - bearerAuth: []
        - oauth2: [categories.write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
      responses:
        '200':
          description: The category as it now stands.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      tags: [Categories]
      operationId: deleteCategory
      summary: Delete a category
      description: >-
        A tombstone rather than a row delete, because the catalogue syncs to the
        phone and a row that simply vanished would come back on the next pull.
        Answers 404 if it was already deleted.
      security:
        - bearerAuth: []
        - oauth2: [categories.write]
      responses:
        '204':
          description: Deleted.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ────────────────────────────────────────────────────────────── oauth ──
...
  /oauth/authorize:
    get:
      tags: [OAuth]
      operationId: authorize
      summary: Start the authorization code flow
      description: |
        Checks the *shape* of the request and redirects the browser to this
        deployment's own consent screen, carrying the parameters along. It never
        redirects to a caller-supplied address and never touches the database,
        which is why it cannot be used to probe whether a client id exists —
        whether the client and the redirect are real is decided later, under the
        signed-in user's session.

        Errors are returned as an RFC 6749 JSON body rather than bounced to
        `redirect_uri`: a validator that redirected errors to an unverified
        address would be an open redirect wearing an OAuth costume.
      security: []
      parameters:
        - name: response_type
          in: query
          required: true
          description: Must be `code`.
          schema:
            type: string
            enum: [code]
        - name: client_id
          in: query
          required: true
          description: The application's public client id.
          schema:
            type: string
            examples: ['wavs_app_5f2c1d9a4b7e42f1a0c6d3e8b91f7a24']
        - name: redirect_uri
          in: query
          required: true
          description: >-
            Where to deliver the code. Must be one of the addresses registered
            for this application — checked later, when the user approves.
          schema:
            type: string
            format: uri
        - name: scope
          in: query
          required: true
          description: >-
            Space-separated. Every entry must be a known scope; one unknown
            entry fails the whole request rather than being silently dropped.
          schema:
            type: string
            examples: ['identity.read groups.read expenses.write offline_access']
        - name: code_challenge
          in: query
          required: true
          description: The unpadded base64url SHA-256 of your verifier — 43 characters.
          schema:
            type: string
            pattern: '^[A-Za-z0-9_-]{43}$'
        - name: code_challenge_method
          in: query
          required: true
          description: '`S256` only. `plain` is a challenge whose answer is in the same request.'
          schema:
            type: string
            enum: [S256]
        - name: state
          in: query
          required: false
          description: Returned unchanged on the redirect back.
          schema:
            type: string
      responses:
        '302':
          description: >-
            Redirect to this deployment's consent screen. The user approves
            there, and the screen calls `POST /oauth/authorize` with their own
            session.
          headers:
            Location:
              description: The consent screen, with the request's parameters attached.
              schema:
                type: string
                format: uri
        '400':
          description: >-
            The request is malformed, or `scope` names something this deployment
            does not know.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthError' }
        '500':
          description: >-
            This deployment has no consent screen configured
            (`WAVES_API_WEB_URL` is unset).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthError' }
    post:
      tags: [OAuth]
      operationId: approveAuthorization
      summary: Record the user's approval and mint a code
      description: >-
        Called by the consent screen, authenticated by the *user's own Waves
        session* rather than a client credential — the decision being recorded
        is theirs. The redirect and the scopes are re-checked against the
        application's registration here, because the browser between the two
        screens is the attacker's ground and nothing it carries is believed
        twice. Returns the address to send the browser to, rather than
        redirecting, so the screen can control the navigation.


        Because it carries a session, this endpoint is on the console's CORS
        surface, not the public one: a browser can only call it from an origin
        in the deployment's `WAVES_API_ALLOWED_ORIGINS` allowlist, and a
        deployment that has not set that variable refuses browser calls
        entirely. The `GET` beside it carries no credential and stays public.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveAuthorizationRequest'
      responses:
        '200':
          description: Where to send the browser, with the code attached.
          content:
            application/json:
              schema:
                type: object
                required: [redirect_to]
                properties:
                  redirect_to:
                    type: string
                    format: uri
                    description: >-
                      The application's registered redirect with `code` and, if
                      you sent one, `state` appended.
        '400':
          description: Malformed, or the redirect or scopes are not registered for this client.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthError' }
        '401':
          description: No Waves session, or one the authorization server would not vouch for.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthError' }
        '500':
          description: Something went wrong on our side.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthError' }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: wavs_pat_… or wavs_at_…
      description: |
        A Waves API token in `Authorization: Bearer …`. Two kinds are accepted
        here and they behave identically once presented:

        - `wavs_pat_…` — a personal access token, created in the developer
          console, acting for the developer's own account.
        - `wavs_at_…` — an OAuth access token, acting for a user who approved an
          application.

        A refresh token (`wavs_rt_…`) is a credential for `/oauth/token` and
        nothing else; presenting one here is refused with `invalid_token`.

        The token string carries no scope list — the row does — so the scopes
        each operation lists apply to both kinds, and narrowing an
        application's scopes takes effect on its next request.

        A missing token is `401 unauthorized` with
        `WWW-Authenticate: Bearer realm="waves"`. A token that is malformed,
        expired, revoked, or belongs to a disabled application is `401
        invalid_token` with
        `WWW-Authenticate: Bearer realm="waves", error="invalid_token"`. A live
        token that lacks the scope the route needs is `403 insufficient_scope`.
    oauth2:
      type: oauth2
      description: >-
        OAuth 2.1 authorization code with PKCE (`S256` only). The URLs below are
        relative to this deployment's own base URL; the authoritative copies are
        in `/.well-known/oauth-authorization-server`, which derives them from
        the request. Ask for `offline_access` if you need a refresh token.
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          refreshUrl: /oauth/token
          scopes:
            identity.read: See your name, avatar, country, language, default currency and payment
...
  headers:
...
    RateLimitRemaining:
      description: Requests left in the current window.
      schema:
        type: integer
...
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: >-
        How many rows to return. Defaults to 25; anything above 100 is quietly
        capped at 100. A value below 1, or one that is not a whole number, is a
        400.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Cursor:
      name: cursor
      in: query
      required: false
      description: >-
        The `next_cursor` from the previous page. Opaque: it encodes a sort key
        and a tiebreaker, and treating it as anything other than a string you
        pass back will break. A cursor this deployment did not issue is a 400.
      schema:
        type: string
        examples:
          ['MjAyNi0wOS0wOFQxMDoxNDowMi4xMTFaIDNmYTg1ZjY0LTU3MTctNDU2Mi1iM2ZjLTJjOTYzZjY2YWZhNg']
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: >-
        A key of your own choosing, 8 to 200 characters, that makes this write
        safe to retry: the same key on the same route returns the original write
        rather than making a second one. (`POST /v1/categories` is the one
        exception — a retry there is refused with a 409 instead of returning the
        original row. Either way you never get a duplicate.) It is scoped to the
        token that sent it, so two developers who both pick `"1"` do not
        collide. Omitting it means the write is not idempotent, which is a
        choice, not an error — but a key outside 8–200 characters is a 400 with
        `details.header = "Idempotency-Key"`.
      schema:
        type: string
        minLength: 8
        maxLength: 200
        examples: ['dinner-2026-09-08-goa-01']
...
  responses:
    BadRequest:
      description: >-
        `invalid_request` — the body, a query parameter or a header is not
        something this endpoint can act on. The message names the field.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: >-
        `unauthorized` when no credential was sent, or `invalid_token` when one
        was sent and is malformed, expired, revoked, of the wrong kind, or
        belongs to a disabled application.
      headers:
        WWW-Authenticate: { $ref: '#/components/headers/WWWAuthenticate' }
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: >-
        `insufficient_scope` when the token is live but was not granted the
        scope this route needs, or `forbidden` when the account itself may not
        do this — not a member of the group, not a party to the settlement, not
        the owner, or a guest whose trial has ended.
      headers:
        WWW-Authenticate: { $ref: '#/components/headers/WWWAuthenticate' }
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
        RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
        RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
        RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Conflict:
      description: >-
        `conflict` — the write collided with the state it was based on. A stale
        `if_revision` or `base_version_no`, shares that do not reconcile with
        the amount, or an id that is already taken.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
        RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
        RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
        RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    TooManyRequests:
      description: >-
        `rate_limited`. Two buckets are enforced, per token and per account, and
        this is returned when either is spent. Wait `Retry-After` seconds.
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
        RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
        RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
        RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InternalError:
      description: >-
        `internal` when something went wrong on our side, or `misconfigured`
        when the deployment is missing configuration it needs (check `/health`).
        Neither says anything about the cause; quote `request_id`.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }

  # ──────────────────────────────────────────────────────────── schemas ──
  schemas:
...
    QueryFlag:
      type: string
      enum: ['true', '1', 'yes', 'on', 'false', '0', 'no', 'off']
      description: >-
        A boolean in a query string, matched case-insensitively. `true`, `1`,
        `yes` and `on` are true; `false`, `0`, `no` and `off` are false; an
        empty or absent value is false. Anything else is a 400 with
        `details.parameter` naming the offender — a strict `=== "true"` would
        read `?flag=1` as false without saying so, which is the worst kind of
        refusal because the caller sees a plausible answer to a different
        question.
...
    GroupPage:
      type: object
      required: [data, next_cursor]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Group'
        next_cursor:
          type: [string, 'null']
          description: >-
            Pass as `cursor` for the next page, or null when this is the last
            one. A page can be shorter than `limit` and still have one.
...